#255 - extract common code to Utils
