+++ src/common/TransactionParser.ts
@@ -1,8 +1,7 @@
 import * as winston from "winston";
 import { Transaction } from "../models/TransactionModel";
 import { TransactionOperation } from "../models/TransactionOperationModel";
-import { removeScientificNotationFromNumbString } from "./Utils";
-import { Config } from "./Config";
+import { removeScientificNotationFromNumbString, fetchTransactionReceipts } from "./Utils";
 import * as Bluebird from "bluebird";
 import { IDecodedLog, IContract, ITransaction, IBlock, IExtractedTransaction, ISavedTransaction } from "./CommonInterfaces";
 const erc20abi = require("./contracts/Erc20Abi");
@@ -153,50 +152,8 @@ export class TransactionParser {
         })
     }
 
-    private async fetchTransactionReceipts (transactions: any) {
-        const batchLimit = 300
-        const chunk = (list, size) => list.reduce((r, v) =>
-            (!r.length || r[r.length - 1].length === size ?
-            r.push([v]) : r[r.length - 1].push(v)) && r
-        , []);
-        const chunkTransactions = chunk(transactions, batchLimit)
-
-        try {
-            const receipts = await Bluebird.map(chunkTransactions, (chunk: any) => {
-                return new Promise((resolve, reject) => {
-                  let completed = false;
-                  const chunkReceipts = [];
-                  const callback = (err: Error, receipt: any) => {
-                    if (completed) return;
-                    if (err || !receipt) {
-                        completed = true;
-                        reject(err);
-                    }
-
-                    chunkReceipts.push(err ? null : receipt);
-                    if (chunkReceipts.length >= chunk.length) {
-                        completed = true;
-                        resolve(chunkReceipts);
-                    }
-                  };
-
-                  if (chunk.length > 0) {
-                    const batch = new Config.web3.BatchRequest();
-                    chunk.forEach((tx: any) => {
-                        batch.add(Config.web3.eth.getTransactionReceipt.request(tx, callback));
-                    });
-                    batch.execute();
-                  } else {
-                    resolve(chunkReceipts);
-                  }
-                });
-            })
-
-            return [].concat(...receipts)
-        } catch (error) {
-            winston.error(`Error getting transtaction receipt `, error)
-            Promise.reject(error)
-        }
+    private async fetchTransactionReceipts (transactionsIDs: string[]) {
+        return fetchTransactionReceipts(transactionsIDs);
     }
 
     static getTransactions(blockNumber: number): Promise<any[]> {
