+++ src/common/Utils.ts
@@ -1,6 +1,7 @@
 import { Response } from "express";
 import * as winston from "winston";
 import { Config } from "./Config";
+import * as Bluebird from "bluebird";
 const axios = require("axios");
 
 /**
@@ -125,3 +126,49 @@ Object.defineProperty(Array.prototype, "flatMap", {
     },
     enumerable: false,
 })
+
+export async function fetchTransactionReceipts (transactionIDs: string[]) {
+    const batchLimit = 300
+    const chunk = (list, size) => list.reduce((r, v) =>
+        (!r.length || r[r.length - 1].length === size ?
+            r.push([v]) : r[r.length - 1].push(v)) && r
+        , []);
+    const chunkTransactions = chunk(transactionIDs, batchLimit)
+
+    try {
+        const receipts = await Bluebird.map(chunkTransactions, (chunk: any) => {
+            return new Promise((resolve, reject) => {
+                let completed = false;
+                const chunkReceipts = [];
+                const callback = (err: Error, receipt: any) => {
+                    if (completed) return;
+                    if (err || !receipt) {
+                        completed = true;
+                        reject(err);
+                    }
+
+                    chunkReceipts.push(err ? null : receipt);
+                    if (chunkReceipts.length >= chunk.length) {
+                        completed = true;
+                        resolve(chunkReceipts);
+                    }
+                };
+
+                if (chunk.length > 0) {
+                    const batch = new Config.web3.BatchRequest();
+                    chunk.forEach((tx: any) => {
+                        batch.add(Config.web3.eth.getTransactionReceipt.request(tx, callback));
+                    });
+                    batch.execute();
+                } else {
+                    resolve(chunkReceipts);
+                }
+            });
+        })
+
+        return [].concat(...receipts)
+    } catch (error) {
+        winston.error(`Error getting transtaction receipt `, error)
+        Promise.reject(error)
+    }
+}
